bug with overpass server overload

+++ src/fetchActivityLocations.py
@@ -27,37 +27,46 @@ def fetchActivityLocations(latitude, longitude):
     out skel qt; 
     '''
     api = overpy.Overpass()                       # creating a overpass API instance 
-    result = api.query(built_query)               # get result from API by sending the query to overpass servers
+    serverFree = 1
+    try:
+        result = api.query(built_query)               # get result from API by sending the query to overpass servers
+    except overpy.exception.OverpassGatewayTimeout:
+        serverFree = 0
+        print("Overpass server is at capacity! Please try again later")
 
-    list_of_node_tags = []                        # initializing empty list , we'll use it to form a dataframe .
-    for node in result.nodes:                     # from each node , get the all tags information
-        node.tags['latitude'] =  node.lat
-        node.tags['longitude'] = node.lon
-        node.tags['id'] = node.id
-        list_of_node_tags.append(node.tags)
-    data_frame = pd.DataFrame(list_of_node_tags)  # forming a pandas dataframe using list of dictionaries
-    # print(data_frame)
-    for column_name in data_frame:
-        # print(column_name)
-        if not any(x in column_name for x in ("name", "longitude", "latitude")): # check to see if any column names dont contain "name", "longitude", latitude
-            data_frame.drop(column_name, axis=1, inplace=True) # drop the columns if above is true
-    # print(data_frame)
-    #first checks name column if there is one and stores in a list of activity locations and then removes both the column and the rows for optimation purposes 
-    if "name" in data_frame.columns:
-        for row in data_frame.itertuples():
-            # print(row)
-            if str(row.name) != "nan":
-                #creates new activity locaiton object containging name and latitude, longitude information
-                newActivityLocation = ActivityLocation.ActivityLocation(str(row.name), row.latitude, row.longitude)
-                # activityLocationList.append(str(row.name))
-                activityLocationList.append(newActivityLocation) # Add activity location object to list of activity locations
-                activityLocationIndex.append(row.Index) # add index of row to indexList to ensure there are not any duplicates
-                data_frame.drop(row.Index, inplace=True) #DROP COLUMN WITH COLUMN NAME NAME
-        data_frame.drop("name", axis=1, inplace=True)
+    if (serverFree == 1):
+        list_of_node_tags = []                        # initializing empty list , we'll use it to form a dataframe .
+        for node in result.nodes:                     # from each node , get the all tags information
+            node.tags['latitude'] =  node.lat
+            node.tags['longitude'] = node.lon
+            node.tags['id'] = node.id
+            list_of_node_tags.append(node.tags)
+        data_frame = pd.DataFrame(list_of_node_tags)  # forming a pandas dataframe using list of dictionaries
+        # print(data_frame)
+        for column_name in data_frame:
+            # print(column_name)
+            if not any(x in column_name for x in ("name", "longitude", "latitude")): # check to see if any column names dont contain "name", "longitude", latitude
+                data_frame.drop(column_name, axis=1, inplace=True) # drop the columns if above is true
+        # print(data_frame)
+        #first checks name column if there is one and stores in a list of activity locations and then removes both the column and the rows for optimation purposes 
+        if "name" in data_frame.columns:
+            for row in data_frame.itertuples():
+                # print(row)
+                if str(row.name) != "nan":
+                    #creates new activity locaiton object containging name and latitude, longitude information
+                    newActivityLocation = ActivityLocation.ActivityLocation(str(row.name), row.latitude, row.longitude)
+                    # activityLocationList.append(str(row.name))
+                    activityLocationList.append(newActivityLocation) # Add activity location object to list of activity locations
+                    activityLocationIndex.append(row.Index) # add index of row to indexList to ensure there are not any duplicates
+                    data_frame.drop(row.Index, inplace=True) #DROP COLUMN WITH COLUMN NAME NAME
+            data_frame.drop("name", axis=1, inplace=True)
+            # print(activityLocationList)
+            return (newStop, activityLocationList) #returns a tuple containting the stop point and a list of activity locaitons near stop point
+            # print(data_frame)                              # return data frame if you want to use it further in main function.
+    else:
+        return 0
 
-    # print(activityLocationList)
-    return (newStop, activityLocationList) #returns a tuple containting the stop point and a list of activity locaitons near stop point
-    # print(data_frame)                              # return data frame if you want to use it further in main function.
+    
 
 
 def fetchStopAL(list_of_stops):
